product 충돌 수정

+++ backend/src/main/kotlin/com/study/commerce/product/repository/ProductRepository.kt
@@ -3,6 +3,6 @@ package com.study.commerce.product.repository
 import com.study.commerce.product.domain.Product
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ProductRepository:JpaRepository<Product,Int> {
-    fun findByProductId(productId:Int): Product
+interface ProductRepository : JpaRepository<Product, Long> {
+    fun findByProductId(productId: Long): Product
 }
